@@ -5,11 +5,15 @@
   "dependencies": {
     "material-ui": "^0.19.4",
     "p5": "^0.5.16",
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "react-p5-wrapper": "0.0.4",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.17",
     "react-tap-event-plugin": "^3.0.2",
+    "redux": "^3.7.2",
+    "redux-thunk": "^2.2.0",
     "rita": "^1.2.12"
   },
   "scripts": {
